apply style only is userscript installed, nightmode
I changed most of the style's looks to only work if the helper
userscript is also installed. This is to prevent it from crapping out
and displaying a bad experience when they only have the style
installed. If people disable the script, the style automagically turns
off as well, preventing crappy displays.

Also started off the RES Nightmode functionality.